requirements.txt
discord.py==2.4.0
python-dateutil==2.9.0